feat: Add FavoriteService protocol and implementation
This commit adds the `FavoriteService` protocol and its implementation in the `FavoriteManager` class. The protocol defines two methods: `getFavorites` and `toggleFavorite`, which are responsible for retrieving a list of favorite products and toggling the favorite status of a product, respectively. The implementation uses the `ApiServiceProvider` to make the necessary API requests. This addition enhances the functionality of the app by providing the ability to manage favorite products.